I synchronized my version with the up-to-date git as of about 9/6/09.
It now gives the same answers as git on the example files, though I didn't
examine all files and options to fine granularity!

Conflicts:
	include/constr.h
	include/mdrun.h
	include/update.h
	src/gmxlib/enxio.c
	src/kernel/md.c
	src/kernel/steep.c
	src/mdlib/coupling.c
	src/mdlib/mdebin.c
	src/mdlib/sim_util.c
	src/mdlib/update.c

+++ src/gmxlib/matio.c
@@ -51,7 +51,7 @@
 #define round(a) (int)(a+0.5)
 
 static const char mapper[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789!@#$%^&*()-_=+{}|;:',<.>/?";
-#define NMAP strlen(mapper)
+#define NMAP (long int)strlen(mapper)
 
 #define MAX_XPM_LINELENGTH 4096
